Modified CostOptimizationHub service to use a custom client as the service is supported only in us-east-1
